Make bistro toplevel-compatible
Nproc cannot be used within a toplevel because of function
serialization: a function created interactively will not be available
to the toplevel subprocesses.

So we introduce a "toplevel" mode, where Nproc is not used
(Lwt_preemptive.detach is used instead) and only one processor is
available.

If/when multicore is available, we can stop using Nproc and do things
more simply!