ppx: better (but still not perfect) location in antiquotations
When parsing an antiquotation, the (ocaml) expression parser is given
the location of the whole extension. Before this fix, the reported
location was inside the string containing the antiquotation. To have
even better location, the (script) AST should contain location info
from the Lexing module. This should be feasible by using an ocamllex
lexer.