Bistro_app.simple: simplified workflow execution
Provides a function for the common case where we want to run a
workflow and put some of the targets in given locations. The function
wraps the creation of a database and a scheduler, as well as
concurrency aspects.

This function is defined in the new Bistro_app module, itself inside
the new bistro.utils library.