engine/Scheduler: give more flexibility to backends
backends now instantiate a workflow's script themselves (that is
injecting the path and resource values in the text of the
script). This was made necessary for pbs backend, to allow for a
destination outside the _bistro database. Indeed on a cluster node,
it's more efficient to write on a local disk, rather than writing on a
network volume. This is also a first step for backends to choose how
many processors they want to allocate.