Introduction of packages-as-external-dependencies
With this commit, I start removing workflows corresponding to
installation scripts, as this is out-of-scope of the library. To keep
track of programs that should be available for a workflow to be run, I
introduce a `package` type that encodes the name and version of a
necessary program. Values of type `package` can be used in scripts,
thanks to a new token. It is up to each backend to provide the
packages. For now, we just assume `packages` are in the PATH.

Workflows corresponding to installation scripts will progressively be
removed.