Further externalization of packages
Packages are now declared at workflow level, they cannot be used in a
script anymore. This required too much coupling with the backend
providing packages. This backend is now supposed to set everything
right for the *programs* of a package to be run. I started removing
installation scripts, but this requires more attention and should be
dealt with on a case by case basis.