Massive refactoring: introduction of Bistro.Task
The goal here is to decouple the construction of a workflow and its
construction by offering two distinct representations. The first phase
is about manipulating expressions and a tree-like representation seems
a good match. The second is more akin to manipulating a graph of
tasks. This commit introduces the latter representation with the
module named [Bistro.Task], and it adapts the existing code the
database and the scheduler.

The workflow type has changed significantly but the DSL is still
pretty much the same, which means there was only very little
modifications of the the wrappers.