engine/Scheduler: new representation for concrete steps of running a task
Declared a type to represent the concrete (no more symbols) actions to
take for executing a task. This is to decouple the generation of
command lines/scripts from their execution/creation (to stop
generating scripts on disk when interpreting the symbols, which is
very ugly). This should also allow better error reporting.

Also declared a type to represent an execution environment (normal or
docker). Avoided much code duplication with that.