Optimized building of numerous targets
Building many targets at once was slow due to the engine importing the
set of tasks from a workflow separately for each workflow. In
practice, the workflows may share a large part of their tasks, so it
is more efficient to merge task sets in memory and then write it in
the database.