Handling of file dumps in scripts
Using `Bistro.EDSL.dump`, it is now possible to generate
configuration files for wrapped tools in a very easy way: just specify
the contents of the configuration file (which may contain deps) and
create a so-called file dump token, which can be inserted the command
line. Instead of the contents of the configuration file, a path is
inserted and the config file is generated when the workflow is
executed.

In summary, one can now:
- generate config files or scripts using the EDSL
- from which deps are tracked
- without bothering giving a name to this generated file
- in a way that is compatible with docker

:o)