Refactoring to avoid Task.config leaking everywhere
Task now produce a `result` which has to contain everything relevant
about the execution of the task. That way, loggers don't have to
access the config/database. In the process, this commit fixes a
problem with html logger, which would log commands with maximum
resources, and not resources really available during execution.