fixed performance bug in computing ids
Step ids were computed using a simple digest of the command. However,
since commands carry dependencies with them, the digest is computed
from all deps, and the complexity of creating a workflow is then
linear in the number of direct *and* indirect dependencies. Which
means creating n workflows is O(n^2) complexity.

This commit changes id computation so that its complexity is only the
number of direct dependencies, which is bounded (and very low) in
practice.