engine/Scheduler: fixed performance bug in DAG construction
to avoid repeating the construction of a DAG chunk from an already
seen workflow, I would compute the task and check if it was already in
the dag. Unfortunately, this would lead to building a big number of
task representations for nothing, hence lots of useless
allocations. Symptoms were unreasonable starting time before running
the first command and explosion of memory footprint.

I just added a map that keeps already constructed tasks given the
corresponding workflow id, with a huge performance improvement. Now in
the "accordion" test, the DAG for a 100000 step workflow is built in a
few seconds.