tdag: big performance improvement in finding sources of tdag
In order to find sources of tdag, I would use `in_degree` function
which has a terrible complexity in `Graph.Persistent.Concrete`,
since it basically traverses the whole graph to find the degree of a
node. Sigh... Switch to `Graph.Persistent.ConcreteBidirectional`,
which has a far more reasonable complexity.

Now on the "accordion" test, a pipeline with 200000 steps (-n 100000)
starts within 20s on my machine.