engine: live removal of intermediates
This commit leverages the post_revdeps hook introduced in the previous
commit to remove workflow result in cache when we know it's not needed
for any further workflow. Some workflow may be marked as "precious" to
avoid this removal. Worklows mentionned in an App (applicative functor
of computation based on workflows) are marked as precious.

The new engine works as expected, except when we re-run a workflow: it
starts from the beginning.