engine: fixed incorrect handling of precious workflows
There was a bug when a given workflow contained a deps `w` several
times and some occurrences are marked precious while others are
not. In that case, if the first one to be met when constructing the
task graph is not marked, then it is deleted although other
occurrences are marked precious.

This scenario may occur if a workflow is both an intermediate result
and also used in an App.

The proposed fix relies on a function that takes a list of (untyped)
workflows and propagate precious markings to all occurrences of a
workflow.