tdag: explicit build goals
tdag would assume that only sources of the task graph would be build
goals but this is wrong: a workflow can be both an intermediate and a
wanted output. This bug wouldn't show before, because everything was
built and never removed. It's not the case now that we have
incremental removal of unwanted intermediates.