tdag: fixed bug on post-revdeps hook
the post-revdeps hook is called once all revdeps of a task are
finished. However, we should make a distinction on revdeps that
terminated successfully and those which didn't. In bistro use case,
where the post-revdeps hook is used to remove files in the cache, we
definitely want to avoid removing the file if one of the revdeps
failed, since it will be needed when re-running the pipeline after
fixing it.