engine: fixed performance bug related to precious targets
in 00775188a I introduced an operation to propagate "precious"
attributes (if a workflow appears several time in multiple copies, it
may have different values for this attribute). For this I wrote a
recursive function that would first collect all precious workflows
(their id), and rebuild a workflow such that all copies of a
particular step would have "precious" enabled if at least one copy was
enabled in the original version. Sadly, this operation breaks sharing
and could lead to the construction of an unreasonably large workflow
in some real-world cases (see diamond test).

By doing this normalization at a later step (when constructing the
task graph), we encounter no such difficulty and the code is also
easier to follow.