new workflow step map_command
a map_command is a task that has to perform the same command for all
files in a directory in parallel. This is a first implementation
(untested and actually sequential). Ideally, a map command should have
been able to consider whole pipelines for each file of a directory,
but this requires drastic changes to the execution engine.