repo: fixed incorrect relative link computation
when detecting redundancies in repo I needed to test if a path was a
prefix of another, with *directory* granularity. Problem was I tested
that at *character* granularity, causing foo/bar.fa and foo/bar.fa.gz
to be detected as redundant. However for the relevant meaning when
generating repo, only foo is considered redundant.