engine: fixed singularity call
added option -c when invoking singularity exec because the automatic
mount of current directory may sometimes mess up. The example I had
was a ./data link to somewhere outside the current directory, which
was mounted as a dead link, and that would prevent me from binding
things contained in data.