changed granularity of container specification
container specifications in shell workflows changed from being
associated to commands to being associated to the whole workflow. The
motivation for this is that the way shell commands were interpreted
differently when invoked through containers or not, with respect to
quotes. Commands were called with lines like:

  docker run [...] bash -c '<cmd>'

and if <cmd> contained quotes, this resulted in a wrong command
line. Now the commands from a shell workflow are written in a text
file and the container called to run the whole script.

This rules out one possibility that was very seldom used, that is
using different containers in the same script.